Handle more error cases
The encapsulation on this is not the best—AFAIK HTTP 422 for too many
tags is a Danbooru thing and Gelbooru won't do that—but I don't want to
rework how `util.get_json()` works just now.

As stated in the root commit's message, a lot of this probably needs to
be reworked for better abstraction. The goal here is to patch a few
"Unexpected" exceptions that should be handled gracefully just so the
plugins are more stable, while I think about a redesign.