gelbooru: early return for any link that isn't of type `page=post`
As alluded to in the comment, doing this in regex would be prohibitively
complicated, and possibly also unreliable. It's not *that* much overhead
to have Sopel call the plugin function for input it doesn't want. In
fact, this simple substring check is *more* efficient than burning more
cycles in the rule engine processing a complicated regex...