types: unescape HTML entities in tags for display
I don't want to mess around right now with unescaping as the tags get
stored in the object, because that would require that every constructor
do it too (only two now, but could be more later). Since `tag_string()`
is a shared method, it's fine to do it there at the cost of slightly
more work to be done if a single post gets its tags displayed multiple
times (unlikely given that the objects are used once and discarded).